+++ slk/classes/common.py
@@ -1,28 +1,9 @@
 from __future__ import annotations
 
 import binascii
-import sys
 from typing import Any, Type
 
 from xrpl.wallet import Wallet
-
-EPRINT_ENABLED = True
-
-
-def disable_eprint() -> None:
-    global EPRINT_ENABLED
-    EPRINT_ENABLED = False
-
-
-def enable_eprint() -> None:
-    global EPRINT_ENABLED
-    EPRINT_ENABLED = True
-
-
-def eprint(*args: Any, **kwargs: Any) -> None:
-    if not EPRINT_ENABLED:
-        return
-    print(*args, file=sys.stderr, flush=True, **kwargs)
 
 
 class Account:
