+++ slk/sidechain_interaction.py
@@ -23,10 +23,10 @@ from slk.chain.chain import Chain
 from slk.chain.chain_setup import setup_mainchain, setup_sidechain
 from slk.chain.context_managers import sidechain_network, single_node_chain
 from slk.chain.xchain_transfer import main_to_side_transfer, side_to_main_transfer
-from slk.classes.common import disable_eprint, eprint
 from slk.classes.config_file import ConfigFile
 from slk.repl import set_hooks_dir, start_repl
 from slk.sidechain_params import SidechainParams
+from slk.utils.eprint import disable_eprint, eprint
 from slk.utils.log_analyzer import convert_log
